v0.5.9+luau579

+++ luau/CodeGen/src/OptimizeConstProp.cpp
@@ -12,6 +12,7 @@
 #include <vector>
 
 LUAU_FASTINTVARIABLE(LuauCodeGenMinLinearBlockPath, 3)
+LUAU_FASTFLAGVARIABLE(DebugLuauAbortingChecks, false)
 
 namespace Luau
 {
@@ -55,6 +56,12 @@ struct ConstPropState
             return info->tag;
 
         return 0xff;
+    }
+
+    void updateTag(IrOp op, uint8_t tag)
+    {
+        if (RegisterInfo* info = tryGetRegisterInfo(op))
+            info->tag = tag;
     }
 
     void saveTag(IrOp op, uint8_t tag)
@@ -139,8 +146,15 @@ struct ConstPropState
 
     void invalidateRegisterRange(int firstReg, int count)
     {
-        for (int i = firstReg; i < firstReg + count && i <= maxReg; ++i)
-            invalidate(regs[i], /* invalidateTag */ true, /* invalidateValue */ true);
+        if (count == -1)
+        {
+            invalidateRegistersFrom(firstReg);
+        }
+        else
+        {
+            for (int i = firstReg; i < firstReg + count && i <= maxReg; ++i)
+                invalidate(regs[i], /* invalidateTag */ true, /* invalidateValue */ true);
+        }
     }
 
     void invalidateCapturedRegisters()
@@ -202,7 +216,7 @@ struct ConstPropState
         if (RegisterLink* link = instLink.find(instOp.index))
         {
             // Check that the target register hasn't changed the value
-            if (link->version > regs[link->reg].version)
+            if (link->version < regs[link->reg].version)
                 return nullptr;
 
             return link;
@@ -229,9 +243,18 @@ struct ConstPropState
             return;
 
         if (uint32_t* prevIdx = valueMap.find(inst))
-            substitute(function, inst, IrOp{IrOpKind::Inst, *prevIdx});
-        else
-            valueMap[inst] = instIdx;
+        {
+            const IrInst& prev = function.instructions[*prevIdx];
+
+            // Previous load might have been removed as unused
+            if (prev.useCount != 0)
+            {
+                substitute(function, inst, IrOp{IrOpKind::Inst, *prevIdx});
+                return;
+            }
+        }
+
+        valueMap[inst] = instIdx;
     }
 
     // Vm register load can be replaced by a previous load of the same version of the register
@@ -253,23 +276,28 @@ struct ConstPropState
         // Check if there is a value that already has this version of the register
         if (uint32_t* prevIdx = valueMap.find(versionedLoad))
         {
-            // Previous value might not be linked to a register yet
-            // For example, it could be a NEW_TABLE stored into a register and we might need to track guards made with this value
-            if (!instLink.contains(*prevIdx))
-                createRegLink(*prevIdx, loadInst.a);
+            const IrInst& prev = function.instructions[*prevIdx];
 
-            // Substitute load instructon with the previous value
-            substitute(function, loadInst, IrOp{IrOpKind::Inst, *prevIdx});
+            // Previous load might have been removed as unused
+            if (prev.useCount != 0)
+            {
+                // Previous value might not be linked to a register yet
+                // For example, it could be a NEW_TABLE stored into a register and we might need to track guards made with this value
+                if (!instLink.contains(*prevIdx))
+                    createRegLink(*prevIdx, loadInst.a);
+
+                // Substitute load instructon with the previous value
+                substitute(function, loadInst, IrOp{IrOpKind::Inst, *prevIdx});
+                return;
+            }
         }
-        else
-        {
-            uint32_t instIdx = function.getInstIndex(loadInst);
 
-            // Record load of this register version for future substitution
-            valueMap[versionedLoad] = instIdx;
+        uint32_t instIdx = function.getInstIndex(loadInst);
 
-            createRegLink(instIdx, loadInst.a);
-        }
+        // Record load of this register version for future substitution
+        valueMap[versionedLoad] = instIdx;
+
+        createRegLink(instIdx, loadInst.a);
     }
 
     // VM register loads can use the value that was stored in the same Vm register earlier
@@ -449,9 +477,16 @@ static void constPropInInst(ConstPropState& state, IrBuilder& build, IrFunction&
                 }
 
                 if (state.tryGetTag(source) == value)
-                    kill(function, inst);
+                {
+                    if (FFlag::DebugLuauAbortingChecks)
+                        replace(function, block, index, {IrCmd::CHECK_TAG, inst.a, inst.b, build.undef()});
+                    else
+                        kill(function, inst);
+                }
                 else
+                {
                     state.saveTag(source, value);
+                }
             }
             else
             {
@@ -619,13 +654,20 @@ static void constPropInInst(ConstPropState& state, IrBuilder& build, IrFunction&
         if (uint8_t tag = state.tryGetTag(inst.a); tag != 0xff)
         {
             if (tag == b)
-                kill(function, inst);
+            {
+                if (FFlag::DebugLuauAbortingChecks)
+                    replace(function, inst.c, build.undef());
+                else
+                    kill(function, inst);
+            }
             else
+            {
                 replace(function, block, index, {IrCmd::JUMP, inst.c}); // Shows a conflict in assumptions on this path
+            }
         }
         else
         {
-            state.saveTag(inst.a, b); // We can assume the tag value going forward
+            state.updateTag(inst.a, b); // We can assume the tag value going forward
         }
         break;
     }
@@ -633,25 +675,46 @@ static void constPropInInst(ConstPropState& state, IrBuilder& build, IrFunction&
         if (RegisterInfo* info = state.tryGetRegisterInfo(inst.a))
         {
             if (info->knownNotReadonly)
-                kill(function, inst);
+            {
+                if (FFlag::DebugLuauAbortingChecks)
+                    replace(function, inst.b, build.undef());
+                else
+                    kill(function, inst);
+            }
             else
+            {
                 info->knownNotReadonly = true;
+            }
         }
         break;
     case IrCmd::CHECK_NO_METATABLE:
         if (RegisterInfo* info = state.tryGetRegisterInfo(inst.a))
         {
             if (info->knownNoMetatable)
-                kill(function, inst);
+            {
+                if (FFlag::DebugLuauAbortingChecks)
+                    replace(function, inst.b, build.undef());
+                else
+                    kill(function, inst);
+            }
             else
+            {
                 info->knownNoMetatable = true;
+            }
         }
         break;
     case IrCmd::CHECK_SAFE_ENV:
         if (state.inSafeEnv)
-            kill(function, inst);
+        {
+            if (FFlag::DebugLuauAbortingChecks)
+                replace(function, inst.a, build.undef());
+            else
+                kill(function, inst);
+        }
         else
+        {
             state.inSafeEnv = true;
+        }
         break;
     case IrCmd::CHECK_GC:
         // It is enough to perform a GC check once in a block
